+++ feeds/luci/luci-app-mcu-display/htdocs/luci-static/resources/mcu-display-core.js
@@ -130,7 +130,7 @@ return baseclass.extend({
 	statusFingerprint: function(st) {
 		st = st || {};
 		return [
-			st.running, st.fifo_ok, st.boot_stage,
+			st.running,
 			st.page_id, st.page_idx, st.page_title
 		].join('|');
 	},